+++ newrelic/hooks/framework_tornado/__init__.py
@@ -1,3 +1,4 @@
+import warnings
 import logging
 import weakref
 import traceback
@@ -334,3 +335,18 @@ def finalize_request_monitoring(request, exc=None, value=None, tb=None):
         request._nr_transaction = None
         request._nr_wait_function_trace = None
         request._nr_request_finished = True
+
+
+warnings.warn((
+    'All existing Tornado instrumentation and feature flags will be removed '
+    'in a future release of our agent. Only Tornado 6.x and up will be '
+    'supported without feature flags in a future release of the Python '
+    'Agent. For more information see '
+    'https://discuss.newrelic.com/t/python-tornado-support/75613'),
+    DeprecationWarning)
+print(
+'DeprecationWarning: All existing Tornado instrumentation and feature flags '
+'will be removed in a future release of our agent. Only Tornado 6.x and up '
+'will be supported without feature flags in a future release of the Python '
+'Agent. For more information see '
+'https://discuss.newrelic.com/t/python-tornado-support/75613')