+++ newrelic/hooks/framework_tornado_r1.py
@@ -1,3 +1,4 @@
+import warnings
 import logging
 import sys
 import weakref
@@ -1245,3 +1246,18 @@ def instrument_tornado_wsgi(module):
     module.WSGIContainer.__call__ = ObjectWrapper(
             module.WSGIContainer.__call__, None,
             wsgi_container_call_wrapper)
+
+
+warnings.warn((
+    'All existing Tornado instrumentation and feature flags will be removed '
+    'in a future release of our agent. Only Tornado 6.x and up will be '
+    'supported without feature flags in a future release of the Python '
+    'Agent. For more information see '
+    'https://discuss.newrelic.com/t/python-tornado-support/75613'),
+    DeprecationWarning)
+print(
+'DeprecationWarning: All existing Tornado instrumentation and feature flags '
+'will be removed in a future release of our agent. Only Tornado 6.x and up '
+'will be supported without feature flags in a future release of the Python '
+'Agent. For more information see '
+'https://discuss.newrelic.com/t/python-tornado-support/75613')